requirements.txt
exceptiongroup==1.3.1
iniconfig==2.1.0
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.2
PyYAML==6.0.3
tomli==2.4.1
typing_extensions==4.15.0

cma>=3.3.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
mlflow>=2.10.0
rich>=13.0.0